requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3
APScheduler==3.10.4
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
ta==0.10.2
httpx==0.25.2

